requirements.txt
Django==5.1
djangorestframework==3.15.2
uvicorn[standard]==0.21.1
gunicorn==20.1.0
psycopg2-binary==2.9.9


